pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dummy-wx>=0.3.0",
    "numpy>=2.2.4",
    "pillow>=11.1.0",
    "pyopengl>=3.1.9",
    "pyqt5>=5.15.11",
    "requests>=2.32.3",
    "websocket-client>=1.8.0",
]
